CF-466 Refactor parameter's validations for all OS objects.
* Change validation logic to not fail by the first found invalid object - make the list of invalid migrated objects
* Extend flavors parameters to check - move them to different test_file as different test cases for each parameter.
* Don't skip validation if src list is empty - fail it
