+++ cloudferry_devlab/requirements.txt
@@ -11,3 +11,4 @@ nose==1.3.7
 nose-testconfig==0.9.1
 sqlalchemy==1.0.12
 PyMySQL==0.6.7
+git+https://github.com/kevinastone/generator.git@nose-attribs#egg=test-generator
